requirements.txt
Flask==2.3.2
torch
torchvision
pandas
Pillow
